Moved deployer and scheduler into managers
